Move fingerprint calculations from metadata_utils.py to add_fingerprint.py
