+++ matchms/filtering/filter_utils/metadata_utils.py
@@ -1,7 +1,5 @@
 import re
 from typing import Optional
-import numpy as np
-
 
 try:  # rdkit is not included in pip package
     from rdkit import Chem
@@ -147,96 +145,3 @@ def is_valid_inchikey(inchikey: str) -> bool:
         return True
     return False
 
-
-def derive_fingerprint_from_smiles(smiles: str, fingerprint_type: str, nbits: int) -> np.ndarray:
-    """Calculate molecule fingerprint based on given smiles or inchi (using rdkit).
-    Requires conda package *rdkit* to be installed.
-
-    Parameters
-    ----------
-    smiles
-        Input smiles to derive fingerprint from.
-    fingerprint_type
-        Determine method for deriving molecular fingerprints. Supported choices are 'daylight',
-        'morgan1', 'morgan2', 'morgan3'.
-    nbits
-        Dimension or number of bits of generated fingerprint.
-
-    Returns
-    -------
-    fingerprint
-        Molecular fingerprint.
-    """
-    if not _has_rdkit:
-        raise ImportError(rdkit_missing_message)
-
-    mol = Chem.MolFromSmiles(smiles)
-    if mol is None:
-        return None
-    return mol_to_fingerprint(mol, fingerprint_type, nbits)
-
-
-def derive_fingerprint_from_inchi(inchi: str, fingerprint_type: str, nbits: int) -> np.ndarray:
-    """Calculate molecule fingerprint based on given inchi (using rdkit).
-    Requires conda package *rdkit* to be installed.
-
-    Parameters
-    ----------
-    inchi
-        Input InChI to derive fingerprint from.
-    fingerprint_type
-        Determine method for deriving molecular fingerprints. Supported choices are 'daylight',
-        'morgan1', 'morgan2', 'morgan3'.
-    nbits
-        Dimension or number of bits of generated fingerprint.
-
-    Returns
-    -------
-    fingerprint: np.array
-        Molecular fingerprint.
-    """
-    if not _has_rdkit:
-        raise ImportError(rdkit_missing_message)
-
-    mol = Chem.MolFromInchi(inchi)
-    if mol is None:
-        return None
-    return mol_to_fingerprint(mol, fingerprint_type, nbits)
-
-
-def mol_to_fingerprint(mol: Chem.rdchem.Mol, fingerprint_type: str, nbits: int) -> np.ndarray:
-    """Convert rdkit mol (molecule) to molecular fingerprint.
-    Requires conda package *rdkit* to be installed.
-
-    Parameters
-    ----------
-    mol
-        Input rdkit molecule.
-    fingerprint_type
-        Determine method for deriving molecular fingerprints.
-        Supported choices are 'daylight', 'morgan1', 'morgan2', 'morgan3'.
-    nbits
-        Dimension or number of bits of generated fingerprint.
-
-    Returns
-    -------
-    fingerprint
-        Molecular fingerprint.
-    """
-    if not _has_rdkit:
-        raise ImportError(rdkit_missing_message)
-
-    assert fingerprint_type in ["daylight", "morgan1", "morgan2", "morgan3"], "Unkown fingerprint type given."
-
-    if fingerprint_type == "daylight":
-        fp = Chem.RDKFingerprint(mol, fpSize=nbits)
-    elif fingerprint_type == "morgan1":
-        fp = AllChem.GetMorganFingerprintAsBitVect(mol, 1, nBits=nbits)
-    elif fingerprint_type == "morgan2":
-        fp = AllChem.GetMorganFingerprintAsBitVect(mol, 2, nBits=nbits)
-    elif fingerprint_type == "morgan3":
-        fp = AllChem.GetMorganFingerprintAsBitVect(mol, 3, nBits=nbits)
-
-    if fp:
-        return np.array(fp)
-    return None
